fix variable types

+++ include/f/f_pc/f_pc_create_req.h
@@ -25,22 +25,22 @@ typedef struct create_request {
     process_method_tag_class mMtdTg;
     create_request_method_class *mpCtRqMtd;
     void *mpUnk1;
-    int mBsPcId;
+    s32 mBsPcId;
     struct base_process_class *mpRes;
     layer_class *mpLayer;
 } create_request; // Size: 0x48
 
 BEGIN_C_DECLARATIONS
 
-BOOL fpcCtRq_isCreatingByID(create_tag *pTag, int *pId);
-BOOL fpcCtRq_IsCreatingByID(unsigned int id);
+BOOL fpcCtRq_isCreatingByID(create_tag *pTag, s32 *pId);
+BOOL fpcCtRq_IsCreatingByID(u32 id);
 void fpcCtRq_CreateQTo(create_request *pReq);
 void fpcCtRq_ToCreateQ(create_request *pReq);
 BOOL fpcCtRq_Delete(create_request *pReq);
 BOOL fpcCtRq_Cancel(create_request *pReq);
-int fpcCtRq_IsDoing(create_request *pReq);
+s32 fpcCtRq_IsDoing(create_request *pReq);
 void fpcCtRq_Handler(void);
-create_request * fpcCtRq_Create(layer_class *pLayer, unsigned long size, create_request_method_class *pCtRqMtd);
+create_request * fpcCtRq_Create(layer_class *pLayer, u32 size, create_request_method_class *pCtRqMtd);
 
 END_C_DECLARATIONS
